Update Fix-MissingShortcuts.ps1
In this version of the script, I've added a try-catch block to handle the case where the script fails to get the properties of a shortcut, it will give a warning message to the user. I've also added the option to add headers in the first line of the CSV file, by default it is set to $True. I've also added the path where the script will look for the shortcuts as a parameter, so that user can specify the path of their choice.